# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusMarket API",
    "description": "API HTTP/JSON de CampusMarket para crear y consultar publicaciones. El contrato versionado es contracts/openapi-v2.json.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Entorno local"
    }
  ],
  "paths": {
...
    "/usuarios/registro": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Registro",
        "operationId": "registrarUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registro"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Iniciar Sesion",
        "operationId": "iniciarSesion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credenciales"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sesion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/me": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Me",
        "operationId": "consultarUsuarioActual",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "usuarios"
        ],
        "summary": "Perfil",
        "operationId": "editarPerfil",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/usuarios/logout": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Cerrar Sesion",
        "operationId": "cerrarSesion",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Credenciales": {
        "properties": {
          "correo": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Correo"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "correo",
          "password"
        ],
        "title": "Credenciales"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PerfilUpdate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Nombre"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "PerfilUpdate"
      },
...
      "Registro": {
        "properties": {
          "correo": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Correo"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          },
          "nombre": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Nombre"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "correo",
          "password",
          "nombre"
        ],
        "title": "Registro"
      },
      "Sesion": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "expira_en": {
            "type": "string",
            "format": "date-time",
            "title": "Expira En"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expira_en",
          "usuario"
        ],
        "title": "Sesion"
      },
      "Usuario": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "correo": {
            "type": "string",
            "title": "Correo"
          },
          "es_admin": {
            "type": "boolean",
            "title": "Es Admin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "correo",
          "es_admin"
        ],
        "title": "Usuario"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}